requirements.txt
pathspec==0.12.1
PyYAML==6.0.2
